Series of small build and test updates

+++ src/pytest/test_fuzzy.py
@@ -1,4 +1,5 @@
 from rmtest import BaseModuleTestCase
+from redis._compat import long
 import redis
 import unittest
 from hotels import hotels
@@ -17,13 +18,13 @@ class FuzzyTestCase(BaseModuleTestCase):
                                             'body', 'this is a test'))
 
             res = r.execute_command('ft.search', 'idx', '%word%')
-            self.assertEqual(res, [1L, 'doc1', ['title', 'hello world', 'body', 'this is a test']])
+            self.assertEqual(res, [long(1), 'doc1', ['title', 'hello world', 'body', 'this is a test']])
     
     def testLdLimit(self):
         self.cmd('ft.create', 'idx', 'schema', 'title', 'text', 'body', 'text')
         self.cmd('ft.add', 'idx', 'doc1', 1.0, 'fields', 'title', 'hello world')
-        self.assertEqual([1L, 'doc1', ['title', 'hello world']], self.cmd('ft.search', 'idx', '%word%'))  # should be ok
-        self.assertEqual([0L], self.cmd('ft.search', 'idx', r'%sword%'))  # should return nothing
+        self.assertEqual([long(1), 'doc1', ['title', 'hello world']], self.cmd('ft.search', 'idx', '%word%'))  # should be ok
+        self.assertEqual([long(0)], self.cmd('ft.search', 'idx', r'%sword%'))  # should return nothing
     
     def testFuzzyMultipleResults(self):
         with self.redis() as r:
@@ -45,7 +46,7 @@ class FuzzyTestCase(BaseModuleTestCase):
 
 
             res = r.execute_command('ft.search', 'idx', '%word%')
-            self.assertEqual(res, [3L, 'doc3', ['title', 'hello ward', 'body', 'this is a test'], 'doc2', ['title', 'hello word', 'body', 'this is a test'], 'doc1', ['title', 'hello world', 'body', 'this is a test']])
+            self.assertEqual(res, [long(3), 'doc3', ['title', 'hello ward', 'body', 'this is a test'], 'doc2', ['title', 'hello word', 'body', 'this is a test'], 'doc1', ['title', 'hello world', 'body', 'this is a test']])
 
     
     def testFuzzySyntaxError(self):
